online-class/pyproject.toml
[tool.poetry]
name = "online-class"
version = "0.1.0"
description = ""
authors = ["Hamid Nisar <hamidnisarpiaic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.1.1"
httpx = "^0.27.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
